Add files via upload
Tic tac Toe is a classic two player game that involves placing "X" or "O" marks on a 3X3 grid.The goal of the game is to place three of your marks in a row either horizontally,vertically or diagonally.In this project ,we will use python to create a simple command-line version of TIC-TAC_TOE.
We create a function to display the current state of the game board, which will be called after each player's turn. This function will print out the current board with the symbols in their respective positions.
After each player has made their move, you will need to check if there is a winner. This can be done by checking all possible winning combinations, i.e., three in a row horizontally, vertically, or diagonally. If a player has won, the game will end and the winner will be announced. If there is no winner, the game will  considered as draw game or board is full.